Reorganize local functions per R2024a
Starting in R2024a, local functions can be declared anywhere in a MATLAB
script. These versions of the scripts take advantage of that to keep the
function definitions close to where they are initially built and used.

This version also ensures alignment between student version and solution
version for each script.